+++ src/compas_rhino/geometry/brep/brep.py
@@ -28,8 +28,6 @@ from .face import RhinoBrepFace
 from .loop import RhinoBrepLoop
 from .vertex import RhinoBrepVertex
 
-TOLERANCE = 1e-6
-
 
 class RhinoBrep(Brep):
     """Rhino Brep backend class.
@@ -241,7 +239,7 @@ class RhinoBrep(Brep):
             raise BrepError("Failed to create extrusion from curve: {} and vector: {}".format(curve, vector))
         rhino_brep = extrusion.ToBrep()
         if cap_ends:
-            capped = rhino_brep.CapPlanarHoles(TOLERANCE)
+            capped = rhino_brep.CapPlanarHoles(TOL.absolute)
             if capped:
                 rhino_brep = capped
         return cls.from_native(rhino_brep)
@@ -326,7 +324,7 @@ class RhinoBrep(Brep):
             raise BrepError("Cannot check for containment if brep is not manifold or is not closed")
 
         if isinstance(object, Point):
-            return self._brep.IsPointInside(point_to_rhino(object), TOLERANCE, False)
+            return self._brep.IsPointInside(point_to_rhino(object), TOL.absolute, False)
         else:
             raise NotImplementedError
 
@@ -364,55 +362,78 @@ class RhinoBrep(Brep):
         """
         self._brep.Transform(transformation_to_rhino(matrix))
 
-    def trim(self, plane, tolerance=TOLERANCE):
+    def trim(self, plane, tolerance=None):
         """Trim this brep by the given trimming plane.
 
         Parameters
         ----------
         plane : :class:`compas.geometry.Frame` or :class:`compas.geometry.Plane`
             The frame or plane to use when trimming. The discarded bit is in the direction of the frame's normal.
+        tolerance : float, optional
+            The precision to use for the trimming operation. Defaults to `TOL.absolute`.
 
-        tolerance : float
-            The precision to use for the trimming operation.
+        Notes
+        -----
+        Trimming operation may result in multiple results (breps). When trimming, only one is used.
+        The used bit is the one on the opposite side of the cutting plane's normal.
 
         Returns
         -------
         None
 
+        Raises
+        ------
+        BrepTrimmingError
+            If the trimming operation ended with no result.
+
+        See Also
+        --------
+        :meth:`compas_rhino.geometry.RhinoBrep.trimmed`
+
         """
-        results = self.trimmed(plane, tolerance)
-        if not results:
-            raise BrepTrimmingError("Trim operation ended with no result")
+        result = self.trimmed(plane, tolerance)
+        self._brep = result.native_brep
 
-        self._brep = results[0].native_brep
-
-    def trimmed(self, plane, tolerance=TOLERANCE):
+    def trimmed(self, plane, tolerance=None):
         """Returns a trimmed copy of this brep by the given trimming plane.
 
         Parameters
         ----------
         plane : :class:`compas.geometry.Frame` or :class:`compas.geometry.Plane`
             The frame or plane to use when trimming. The discarded bit is in the direction of the plane's normal.
+        tolerance : float, optional
+            The precision to use for the trimming operation. Defaults to `TOL.absolute`.
 
-        tolerance : float
-            The precision to use for the trimming operation.
+        Notes
+        -----
+        Trimming operation may result in multiple results (breps). When trimming, only one is used.
+        The used bit is the one on the opposite side of the cutting plane's normal.
 
         Returns
         -------
         :class:`~compas_rhino.geometry.RhinoBrep`
 
+        Raises
+        ------
+        BrepTrimmingError
+            If the trimming operation ended with no result.
+
+        See Also
+        --------
+        :meth:`compas_rhino.geometry.RhinoBrep.trim`
+
         """
+        tolerance = tolerance or TOL.absolute
         if isinstance(plane, Frame):
             plane = Plane.from_frame(plane)
         results = self._brep.Trim(plane_to_rhino(plane), tolerance)
-
-        breps = []
-        for result in results:
-            capped = result.CapPlanarHoles(TOLERANCE)
-            if capped:
-                result = capped
-            breps.append(RhinoBrep.from_native(result))
-        return breps
+        if not results:
+            raise BrepTrimmingError("Trim operation ended with no result")
+        result = results[0]
+        capped = result.CapPlanarHoles(tolerance)
+        if capped:
+            result = capped
+        return RhinoBrep.from_native(result)
 
     @classmethod
     def from_boolean_difference(cls, breps_a, breps_b):
@@ -438,7 +459,7 @@ class RhinoBrep(Brep):
         resulting_breps = Rhino.Geometry.Brep.CreateBooleanDifference(
             [b.native_brep for b in breps_a],
             [b.native_brep for b in breps_b],
-            TOLERANCE,
+            TOL.absolute,
         )
         return [RhinoBrep.from_native(brep) for brep in resulting_breps]
 
@@ -464,7 +485,7 @@ class RhinoBrep(Brep):
         if not isinstance(breps_b, list):
             breps_b = [breps_b]
 
-        resulting_breps = Rhino.Geometry.Brep.CreateBooleanUnion([b.native_brep for b in breps_a + breps_b], TOLERANCE)
+        resulting_breps = Rhino.Geometry.Brep.CreateBooleanUnion([b.native_brep for b in breps_a + breps_b], TOL.absolute)
         return [RhinoBrep.from_native(brep) for brep in resulting_breps]
 
     @classmethod
@@ -491,7 +512,7 @@ class RhinoBrep(Brep):
         resulting_breps = Rhino.Geometry.Brep.CreateBooleanIntersection(
             [b.native_brep for b in breps_a],
             [b.native_brep for b in breps_b],
-            TOLERANCE,
+            TOL.absolute,
         )
         return [RhinoBrep.from_native(brep) for brep in resulting_breps]
 
@@ -529,5 +550,5 @@ class RhinoBrep(Brep):
             list of zero or more resulting Breps.
 
         """
-        resulting_breps = self._brep.Split(cutter.native_brep, TOLERANCE)
+        resulting_breps = self._brep.Split(cutter.native_brep, TOL.absolute)
         return [RhinoBrep.from_native(brep) for brep in resulting_breps]